+++ docs/apps/schedy/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 
 ### Added
+* Added new `generic2` actor type which is more flexible than the old `generic`.
 
 ### Changed